setup.py: Use gnureadline for all other OS
This adds detection for if the readline library is built in, and use
gnureadline if that is not the case and not on Windows.

Since gnureadline links against a static internal version of the
readline library, it could be used for systems where readline support
is not built in (except on Windows where we have pyreadline).

Closes https://github.com/pombredanne/anyreadline/issues/1